Renaming ATA to IDE and AHCI to SATA
